TIP-607: move structured denormalizers in standard directory
