+++ src/Pim/Component/Catalog/Denormalizer/Standard/ProductValue/AbstractValueDenormalizer.php
@@ -1,0 +1,34 @@
+<?php
+
+namespace Pim\Component\Catalog\Denormalizer\Standard\ProductValue;
+
+use Symfony\Component\Serializer\Normalizer\DenormalizerInterface;
+
+/**
+ * Abstract value denormalizer
+ *
+ * @author    Filips Alpe <filips@akeneo.com>
+ * @copyright 2014 Akeneo SAS (http://www.akeneo.com)
+ * @license   http://opensource.org/licenses/osl-3.0.php  Open Software License (OSL 3.0)
+ */
+abstract class AbstractValueDenormalizer implements DenormalizerInterface
+{
+    /** @var string[] */
+    protected $supportedTypes;
+
+    /**
+     * @param string[] $supportedTypes
+     */
+    public function __construct(array $supportedTypes)
+    {
+        $this->supportedTypes = $supportedTypes;
+    }
+
+    /**
+     * {@inheritdoc}
+     */
+    public function supportsDenormalization($data, $type, $format = null)
+    {
+        return in_array($type, $this->supportedTypes) && 'standard' === $format;
+    }
+}
