+++ src/Pim/Component/ReferenceData/Denormalizer/Standard/ProductValue/ReferenceDataCollectionDenormalizer.php
@@ -1,0 +1,58 @@
+<?php
+
+namespace Pim\Component\ReferenceData\Denormalizer\Standard\ProductValue;
+
+use Doctrine\Common\Collections\ArrayCollection;
+use Pim\Component\Catalog\Denormalizer\Standard\ProductValue\AbstractValueDenormalizer;
+use Symfony\Component\Serializer\Normalizer\DenormalizerInterface;
+
+/**
+ * @author    Adrien Pétremann <adrien.petremann@akeneo.com>
+ * @copyright 2015 Akeneo SAS (http://www.akeneo.com)
+ * @license   http://opensource.org/licenses/osl-3.0.php  Open Software License (OSL 3.0)
+ */
+class ReferenceDataCollectionDenormalizer extends AbstractValueDenormalizer
+{
+    /** @var DenormalizerInterface */
+    protected $referenceDataDenormalizer;
+
+    /**
+     * @param array                 $supportedTypes
+     * @param DenormalizerInterface $referenceDataDenormalizer
+     */
+    public function __construct(
+        array $supportedTypes,
+        DenormalizerInterface $referenceDataDenormalizer
+    ) {
+        parent::__construct($supportedTypes);
+
+        $this->referenceDataDenormalizer = $referenceDataDenormalizer;
+    }
+
+    /**
+     * {@inheritdoc}
+     */
+    public function denormalize($data, $referenceDataClass, $format = null, array $context = [])
+    {
+        if (empty($data)) {
+            return new ArrayCollection();
+        }
+
+        if (false === is_array($data)) {
+            throw new \InvalidArgumentException(sprintf('Data expected to be an array.'));
+        }
+
+        $referenceDataColl = new ArrayCollection();
+
+        foreach ($data as $singleData) {
+            $referenceData = $this->referenceDataDenormalizer
+                ->denormalize($singleData, $referenceDataClass, $format, $context);
+
+            if (null !== $referenceData) {
+                $referenceDataColl->add($referenceData);
+            }
+        }
+
+        return $referenceDataColl;
+    }
+}
